Show users when displaying changelog data

+++ src/Controller/GroupController.php
@@ -15,7 +15,9 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Component\Serializer\Encoder\DecoderInterface;
 
 class GroupController extends AbstractController
@@ -103,6 +105,10 @@ class GroupController extends AbstractController
         GroupRepository $groupRepository,
     ): JsonResponse
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('You don\'t have permission to perform this action!');
+        }
+
         $groupRepository->deleteGroup($group);
 
         return $this->json([], 204);
